rumi

+++ app/Http/Controllers/UserCvController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Aggrement;
 use App\Employee;
 use App\Education;
 use App\EmployeeComputerSkill;
@@ -18,6 +19,7 @@ use App\QuestionObjectiveAns;
 use App\Refree;
 use App\RelativeInCb;
 use App\Traning;
+use App\User;
 use Illuminate\Http\Request;
 use PDF;
 use Auth;
@@ -378,28 +380,69 @@ class UserCvController extends Controller
 //            ->get();
     }
     public function FullCvDelete(Request $r){
+        $empId=$r->id;
+        $check=Jobapply::where('fkemployeeId', $empId)->count();
+        if ($check > 0){
+            return $check;
+        }else{
 
+            RelativeInCb::where('fkemployeeId', $empId)->delete();
+            Refree::where('fkemployeeId', $empId)->delete();
+            Refree::where('fkemployeeId', $empId)->delete();
+            MembershipInSocialNetwork::where('fkemployeeId', $empId)->delete();
+            PreviousWorkInCB::where('fkemployeeId', $empId)->delete();
+            JobExperience::where('fkemployeeId', $empId)->delete();
+            ProfessionalQualification::where('fkemployeeId', $empId)->delete();
+            Traning::where('fkemployeeId', $empId)->delete();
+            EmployeeOtherInfo::where('fk_empId', $empId)->delete();
+            EmpOtherSkill::where('fkemployeeId', $empId)->delete();
+            EmployeeComputerSkill::where('fk_empId', $empId)->delete();
+            EmployeeLanguage::where('fkemployeeId', $empId)->delete();
+            Education::where('fkemployeeId', $empId)->delete();
+            QuestionObjective::where('empId', $empId)->delete();
+            QuestionObjectiveAns::where('fkemployeeId', $empId)->delete();
+            Jobapply::where('fkemployeeId', $empId)->delete();
+            $userId=Employee::find($empId)->fkuserId;
+
+            Employee::destroy($empId);
+            Aggrement::where('fkuserId', $userId)->delete();
+            User::destroy($userId);
+            return 0;
+
+        }
+
+
+
+
+
+    }
+    public function FullCvCompleteDelete(Request $r){
         $empId=$r->id;
 
-        RelativeInCb::where('fkemployeeId', $empId)->delete();
-        Refree::where('fkemployeeId', $empId)->delete();
-        Refree::where('fkemployeeId', $empId)->delete();
-        MembershipInSocialNetwork::where('fkemployeeId', $empId)->delete();
-        PreviousWorkInCB::where('fkemployeeId', $empId)->delete();
-        JobExperience::where('fkemployeeId', $empId)->delete();
-        ProfessionalQualification::where('fkemployeeId', $empId)->delete();
-        Traning::where('fkemployeeId', $empId)->delete();
-        EmployeeOtherInfo::where('fk_empId', $empId)->delete();
-        EmpOtherSkill::where('fkemployeeId', $empId)->delete();
-        EmployeeComputerSkill::where('fk_empId', $empId)->delete();
-        EmployeeLanguage::where('fkemployeeId', $empId)->delete();
-        Education::where('fkemployeeId', $empId)->delete();
-        QuestionObjective::where('empId', $empId)->delete();
-        QuestionObjectiveAns::where('fkemployeeId', $empId)->delete();
-        Jobapply::where('fkemployeeId', $empId)->delete();
-        Employee::destroy($empId);
 
-       // return redirect()->route('cv.admin.manage');
+            RelativeInCb::where('fkemployeeId', $empId)->delete();
+            Refree::where('fkemployeeId', $empId)->delete();
+            Refree::where('fkemployeeId', $empId)->delete();
+            MembershipInSocialNetwork::where('fkemployeeId', $empId)->delete();
+            PreviousWorkInCB::where('fkemployeeId', $empId)->delete();
+            JobExperience::where('fkemployeeId', $empId)->delete();
+            ProfessionalQualification::where('fkemployeeId', $empId)->delete();
+            Traning::where('fkemployeeId', $empId)->delete();
+            EmployeeOtherInfo::where('fk_empId', $empId)->delete();
+            EmpOtherSkill::where('fkemployeeId', $empId)->delete();
+            EmployeeComputerSkill::where('fk_empId', $empId)->delete();
+            EmployeeLanguage::where('fkemployeeId', $empId)->delete();
+            Education::where('fkemployeeId', $empId)->delete();
+            QuestionObjective::where('empId', $empId)->delete();
+            QuestionObjectiveAns::where('fkemployeeId', $empId)->delete();
+            Jobapply::where('fkemployeeId', $empId)->delete();
+            $userId=Employee::find($empId)->fkuserId;
+            Employee::destroy($empId);
+            Aggrement::where('fkuserId', $userId)->delete();
+            User::destroy($userId);
+            return 0;
+
+
     }
 
 }
